Related work items: #1519, #1520, #1578, #1579

+++ Cloudents.Web/Models/WebResponse.cs
@@ -9,9 +9,7 @@ namespace Cloudents.Web.Models
 
         public IList<IFilters> Filters { get; set; }
 
-        public int? Count { get; set; }
-      //  public IEnumerable<KeyValuePair<string, string>> Sort { get; set; }
-
+        public long? Count { get; set; }
         public string NextPageLink { get; set; }
     }
